Cambios en el tipo de datos de las entidades

+++ lib_dominio/Entidades/Compras.cs
@@ -8,7 +8,7 @@ namespace lib_dominio.Entidades
         public int Id { get; set; }
         public DateTime? Fecha { get; set; }
         public string? Codigo { get; set; }
-        public double ValorTotal { get; set; }
+        public decimal ValorTotal { get; set; }
 
         //FK
         public int Clientes { get; set; }
